requirements.txt
fastapi
google-cloud-firestore
uvicorn
python-dotenv
pydantic 